refactored game, added a world struct to store players and enemies, redone game update with dx dy (it also treats diagonal direction speed properly and prints local player differently

+++ internal/utils/utils.go
@@ -4,6 +4,6 @@ import(
 	"golang.org/x/image/math/fixed"
 )
 
-func FixedToFloat(f fixed.Int26_6) float64 {
-	return float64(f) / 64
+func FixedToFloat(f fixed.Int26_6) float32 {
+	return float32(f) / 64
 }
